Did some more refactoring work Still need to separate RPS from base Also need to do some tiny changes
