+++ OpModes/motionLibrary/Comparator.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.OpModes.motionLibrary;
+
+//a enum for compound polls, which specifies comparison mode
+public enum Comparator {
+  AND,
+  OR
+}
